backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Redis
redis==5.0.1
hiredis==2.2.3

# MongoDB
pymongo==4.6.0

# Job queue (optional - can use RQ or Celery)
rq==1.15.1

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Excel file processing
openpyxl>=3.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Environment variables
python-dotenv==1.0.0
